requirements.txt
asgiref==3.5.2
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
PyJWT==2.5.0
pystache==0.6.0
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3
static==1.1.1
tox==3.21.4
pytest==7.1.3
pytest-django==4.5.2
pytest-mock==3.9.0
pytest-factoryboy==2.5.0
factory-boy==2.11.1